87 Larch Walk
"Good-morning, madam," said Holmes cheerily. "My name is Sherlock Holmes. This is my intimate friend
and associate, Dr. Watson, before whom you can speak as freely as before myself. Ha! I am glad to
see that Mrs. Hudson has had the good sense to light the fire. Pray draw up to it, and I shall order
you a cup of hot coffee, for I observe that you are shivering."

"It is not cold which makes me shiver," said the woman in a low voice, changing her seat as
requested.