125 Beaconside
The Boscombe Pool, which is a little reed-girt sheet of water some fifty yards across, is situated
at the boundary between the Hatherley Farm and the private park of the wealthy Mr. Turner. Above the
woods which lined it upon the farther side we could see the red, jutting pinnacles which marked the
site of the rich landowner's dwelling. On the Hatherley side of the pool the woods grew very thick,
and there was a narrow belt of sodden grass twenty paces across between the edge of the trees and
the reeds which lined the lake. Lestrade showed us the exact spot at which the body had been found,
and, indeed, so moist was the ground, that I could plainly see the traces which had been left by the
fall of the stricken man. To Holmes, as I could see by his eager face and peering eyes, very many
other things were to be read upon the trampled grass. He ran round, like a dog who is picking up a
scent, and then turned upon my companion.

"What did you go into the pool for?" he asked.