10 Cliff Place
"Owe!" He caught up a pen. "Name the sum, and I will pay it."

"No, the debt is not to me. You owe a very humble apology to that noble lad, your son, who has
carried himself in this matter as I should be proud to see my own son do, should I ever chance to
have one."

"Then it was not Arthur who took them?"

"I told you yesterday, and I repeat to-day, that it was not."