106 Badgers Dene
It was late before Sherlock Holmes returned. He came back alone, for Lestrade was staying in
lodgings in the town.

"The glass still keeps very high," he remarked as he sat down. "It is of importance that it should
not rain before we are able to go over the ground. On the other hand, a man should be at his very
best and keenest for such nice work as that, and I did not wish to do it when fagged by a long
journey. I have seen young McCarthy."

"And what did you learn from him?"

"Nothing."