185 Beaconside
"For heaven's sake, tell me, then, what is this extraordinary mystery!"

"I will do so, and I will show you the steps by which I reached it. And let me say to you, first,
that which it is hardest for me to say and for you to hear: there has been an understanding between
Sir George Burnwell and your niece Mary. They have now fled together."