68 Balcombe Close
"Is Briony Lodge, Serpentine Avenue, St. John's Wood."

Holmes took a note of it. "One other question," said he. "Was the photograph a cabinet?"

"It was."

"Then, good-night, your Majesty, and I trust that we shall soon have some good news for you. And
good-night, Watson," he added, as the wheels of the royal brougham rolled down the street. "If you
will be good enough to call to-morrow afternoon at three o'clock I should like to chat this little
matter over with you."