113 Harvey Street
"One moment," said Holmes, "are you sure about this whistle and metallic sound? Could you swear to
it?"

"That was what the county coroner asked me at the inquiry. It is my strong impression that I heard
it, and yet, among the crash of the gale and the creaking of an old house, I may possibly have been
deceived."

"Was your sister dressed?"

"No, she was in her night-dress. In her right hand was found the charred stump of a match, and in
her left a match-box."

"Showing that she had struck a light and looked about her when the alarm took place. That is
important. And what conclusions did the coroner come to?"

"He investigated the case with great care, for Dr. Roylott's conduct had long been notorious in the
county, but he was unable to find any satisfactory cause of death. My evidence showed that the door
had been fastened upon the inner side, and the windows were blocked by old-fashioned shutters with
broad iron bars, which were secured every night. The walls were carefully sounded, and were shown to
be quite solid all round, and the flooring was also thoroughly examined, with the same result. The
chimney is wide, but is barred up by four large staples. It is certain, therefore, that my sister
was quite alone when she met her end. Besides, there were no marks of any violence upon her."