158 Bowler Street
He took an orange from the cupboard, and tearing it to pieces he squeezed out the pips upon the
table. Of these he took five and thrust them into an envelope. On the inside of the flap he wrote
"S. H. for J. O." Then he sealed it and addressed it to "Captain James Calhoun, Barque Lone Star,
Savannah, Georgia."

"That will await him when he enters port," said he, chuckling. "It may give him a sleepless night.
He will find it as sure a precursor of his fate as Openshaw did before him."

"And who is this Captain Calhoun?"

"The leader of the gang. I shall have the others, but he first."

"How did you trace it, then?"

He took a large sheet of paper from his pocket, all covered with dates and names.