58 Corbet Close
"To the Leadenhall Street Post Office, to be left till called for. He said that if they were sent to
the office he would be chaffed by all the other clerks about having letters from a lady, so I
offered to typewrite them, like he did his, but he wouldn't have that, for he said that when I wrote
them they seemed to come from me, but when they were typewritten he always felt that the machine had
come between us. That will just show you how fond he was of me, Mr. Holmes, and the little things
that he would think of."

"It was most suggestive," said Holmes. "It has long been an axiom of mine that the little things are
infinitely the most important.  Can you remember any other little things about Mr. Hosmer Angel?"

"He was a very shy man, Mr. Holmes. He would rather walk with me in the evening than in the
daylight, for he said that he hated to be conspicuous. Very retiring and gentlemanly he was. Even
his voice was gentle. He'd had the quinsy and swollen glands when he was young, he told me, and it
had left him with a weak throat, and a hesitating, whispering fashion of speech. He was always well
dressed, very neat and plain, but his eyes were weak, just as mine are, and he wore tinted glasses
against the glare."

"Well, and what happened when Mr. Windibank, your stepfather, returned to France?"

"Mr. Hosmer Angel came to the house again and proposed that we should marry before father came back.
He was in dreadful earnest and made me swear, with my hands on the Testament, that whatever happened
I would always be true to him. Mother said he was quite right to make me swear, and that it was a
sign of his passion. Mother was all in his favour from the first and was even fonder of him than I
was. Then, when they talked of marrying within the week, I began to ask about father; but they both
said never to mind about father, but just to tell him afterwards, and mother said she would make it
all right with him. I didn't quite like that, Mr. Holmes. It seemed funny that I should ask his
leave, as he was only a few years older than me; but I didn't want to do anything on the sly, so I
wrote to father at Bordeaux, where the company has its French offices, but the letter came back to
me on the very morning of the wedding."

"It missed him, then?"