40 Tamworth Drive
"'I should be happy to make myself useful.'

"'Quite so. In dress now, for example. We are faddy people, you know--faddy but kind-hearted. If you
were asked to wear any dress which we might give you, you would not object to our little whim. Heh?'

"'No,' said I, considerably astonished at his words.

"'Or to sit here, or sit there, that would not be offensive to you?'

"'Oh, no.'