47 Down Close
"'A marriage has been arranged [it says] and will, if rumour is correct, very shortly take place,
between Lord Robert St. Simon, second son of the Duke of Balmoral, and Miss Hatty Doran, the only
daughter of Aloysius Doran. Esq., of San Francisco, Cal., U.S.A.'

That is all."