76 Wantage Close
"At the same time," he remarked after a pause, during which he had sat puffing at his long pipe and
gazing down into the fire, "you can hardly be open to a charge of sensationalism, for out of these
cases which you have been so kind as to interest yourself in, a fair proportion do not treat of
crime, in its legal sense, at all. The small matter in which I endeavoured to help the King of
Bohemia, the singular experience of Miss Mary Sutherland, the problem connected with the man with
the twisted lip, and the incident of the noble bachelor, were all matters which are outside the pale
of the law. But in avoiding the sensational, I fear that you may have bordered on the trivial."

"The end may have been so," I answered, "but the methods I hold to have been novel and of interest."

"Pshaw, my dear fellow, what do the public, the great unobservant public, who could hardly tell a
weaver by his tooth or a compositor by his left thumb, care about the finer shades of analysis and
deduction! But, indeed, if you are trivial, I cannot blame you, for the days of the great cases are
past. Man, or at least criminal man, has lost all enterprise and originality. As to my own little
practice, it seems to be degenerating into an agency for recovering lost lead pencils and giving
advice to young ladies from boarding-schools. I think that I have touched bottom at last, however.
This note I had this morning marks my zero-point, I fancy. Read it!" He tossed a crumpled letter
across to me.