64 Ventnor Terrace
"But I am all in the dark."

"Of course you are. You'll know all about it presently. Jump up here. All right, John; we shall not
need you. Here's half a crown. Look out for me to-morrow, about eleven. Give her her head. So long,
then!"

He flicked the horse with his whip, and we dashed away through the endless succession of sombre and
deserted streets, which widened gradually, until we were flying across a broad balustraded bridge,
with the murky river flowing sluggishly beneath us. Beyond lay another dull wilderness of bricks and
mortar, its silence broken only by the heavy, regular footfall of the policeman, or the songs and
shouts of some belated party of revellers. A dull wrack was drifting slowly across the sky, and a
star or two twinkled dimly here and there through the rifts of the clouds. Holmes drove in silence,
with his head sunk upon his breast, and the air of a man who is lost in thought, while I sat beside
him, curious to learn what this new quest might be which seemed to tax his powers so sorely, and yet
afraid to break in upon the current of his thoughts. We had driven several miles, and were beginning
to get to the fringe of the belt of suburban villas, when he shook himself, shrugged his shoulders,
and lit up his pipe with the air of a man who has satisfied himself that he is acting for the best.

"You have a grand gift of silence, Watson," said he. "It makes you quite invaluable as a companion.
'Pon my word, it is a great thing for me to have someone to talk to, for my own thoughts are not
over-pleasant. I was wondering what I should say to this dear little woman to-night when she meets
me at the door."