10 Sunningdale Drive
"Then I suggest that we turn our dinner into a supper and follow up this clue while it is still
hot."

"By all means."

It was a bitter night, so we drew on our ulsters and wrapped cravats about our throats. Outside, the
stars were shining coldly in a cloudless sky, and the breath of the passers-by blew out into smoke
like so many pistol shots. Our footfalls rang out crisply and loudly as we swung through the
doctors' quarter, Wimpole Street, Harley Street, and so through Wigmore Street into Oxford Street.
In a quarter of an hour we were in Bloomsbury at the Alpha Inn, which is a small public-house at the
corner of one of the streets which runs down into Holborn. Holmes pushed open the door of the
private bar and ordered two glasses of beer from the ruddy-faced, white-aproned landlord.