96 Wilson Gardens
"In the dress is a pocket. In the pocket is a card-case. In the card-case is a note. And here is the
very note." He slapped it down upon the table in front of him. "Listen to this:

"'You will see me when all is ready. Come at once. "'F.H.M.'

Now my theory all along has been that Lady St. Simon was decoyed away by Flora Millar, and that she,
with confederates, no doubt, was responsible for her disappearance. Here, signed with her initials,
is the very note which was no doubt quietly slipped into her hand at the door and which lured her
within their reach."

"Very good, Lestrade," said Holmes, laughing. "You really are very fine indeed. Let me see it." He
took up the paper in a listless way, but his attention instantly became riveted, and he gave a
little cry of satisfaction. "This is indeed important," said he.

"Ha! you find it so?"

"Extremely so. I congratulate you warmly."