34 Wantage Close
"'Put the papers on the sundial,' I read, peeping over his shoulder.

"'What papers? What sundial?' he asked.

"'The sundial in the garden. There is no other,' said I; 'but the papers must be those that are
destroyed.'

"'Pooh!' said he, gripping hard at his courage. 'We are in a civilised land here, and we can't have
tomfoolery of this kind. Where does the thing come from?'