62 Holmefield Avenue
The prisoner turned with the reckless air of a man who abandons himself to his destiny. "Be it so,"
said he. "And pray what am I charged with?"

"With making away with Mr. Neville St.--Oh, come, you can't be charged with that unless they make a
case of attempted suicide of it," said the inspector with a grin. "Well, I have been twenty-seven
years in the force, but this really takes the cake."

"If I am Mr. Neville St. Clair, then it is obvious that no crime has been committed, and that,
therefore, I am illegally detained."

"No crime, but a very great error has been committed," said Holmes. "You would have done better to
have trusted your wife."

"It was not the wife; it was the children," groaned the prisoner. "God help me, I would not have
them ashamed of their father. My God! What an exposure! What can I do?"