75 Moor End
"Well, it is a possible supposition."

"You think so, too?"

"I did not say a probable one. But you do not yourself look upon this as likely?"

"I do not think Flora would hurt a fly."

"Still, jealousy is a strange transformer of characters. Pray what is your own theory as to what
took place?"