98 Sunningdale Drive
"And your mother is alive?"

"Oh, yes, mother is alive and well. I wasn't best pleased, Mr. Holmes, when she married again so
soon after father's death, and a man who was nearly fifteen years younger than herself. Father was a
plumber in the Tottenham Court Road, and he left a tidy business behind him, which mother carried on
with Mr. Hardy, the foreman; but when Mr. Windibank came he made her sell the business, for he was
very superior, being a traveller in wines. They got £4700 for the goodwill and interest, which
wasn't near as much as father could have got if he had been alive."

I had expected to see Sherlock Holmes impatient under this rambling and inconsequential narrative,
but, on the contrary, he had listened with the greatest concentration of attention.