72 Wilson Gardens
"On June 3rd, that is, on Monday last, McCarthy left his house at Hatherley about three in the
afternoon and walked down to the Boscombe Pool, which is a small lake formed by the spreading out of
the stream which runs down the Boscombe Valley. He had been out with his serving-man in the morning
at Ross, and he had told the man that he must hurry, as he had an appointment of importance to keep
at three. From that appointment he never came back alive.

"From Hatherley Farm-house to the Boscombe Pool is a quarter of a mile, and two people saw him as he
passed over this ground. One was an old woman, whose name is not mentioned, and the other was
William Crowder, a game-keeper in the employ of Mr. Turner. Both these witnesses depose that Mr.
McCarthy was walking alone. The game-keeper adds that within a few minutes of his seeing Mr.
McCarthy pass he had seen his son, Mr. James McCarthy, going the same way with a gun under his arm.
To the best of his belief, the father was actually in sight at the time, and the son was following
him. He thought no more of the matter until he heard in the evening of the tragedy that had
occurred.