141 Holmefield Avenue
"Half dazed, I went into the station and asked about the morning train. There would be one to
Reading in less than an hour. The same porter was on duty, I found, as had been there when I
arrived. I inquired of him whether he had ever heard of Colonel Lysander Stark. The name was strange
to him. Had he observed a carriage the night before waiting for me? No, he had not. Was there a
police-station anywhere near? There was one about three miles off.

"It was too far for me to go, weak and ill as I was. I determined to wait until I got back to town
before telling my story to the police. It was a little past six when I arrived, so I went first to
have my wound dressed, and then the doctor was kind enough to bring me along here. I put the case
into your hands and shall do exactly what you advise."

We both sat in silence for some little time after listening to this extraordinary narrative. Then
Sherlock Holmes pulled down from the shelf one of the ponderous commonplace books in which he placed
his cuttings.

"Here is an advertisement which will interest you," said he. "It appeared in all the papers about a
year ago. Listen to this: