74 Badgers Dene
"His face fell immediately.

"'Dear me!' he said gravely, 'that is very serious indeed! I am sorry to hear you say that. The fund
was, of course, for the propagation and spread of the red-heads as well as for their maintenance. It
is exceedingly unfortunate that you should be a bachelor.'