90 Badgers Dene
"Dr. Becher's."

"Tell me," broke in the engineer, "is Dr. Becher a German, very thin, with a long, sharp nose?"

The station-master laughed heartily. "No, sir, Dr. Becher is an Englishman, and there isn't a man in
the parish who has a better-lined waistcoat. But he has a gentleman staying with him, a patient, as
I understand, who is a foreigner, and he looks as if a little good Berkshire beef would do him no
harm."

The station-master had not finished his speech before we were all hastening in the direction of the
fire. The road topped a low hill, and there was a great widespread whitewashed building in front of
us, spouting fire at every chink and window, while in the garden in front three fire-engines were
vainly striving to keep the flames under.

"That's it!" cried Hatherley, in intense excitement. "There is the gravel-drive, and there are the
rose-bushes where I lay. That second window is the one that I jumped from."

"Well, at least," said Holmes, "you have had your revenge upon them. There can be no question that
it was your oil-lamp which, when it was crushed in the press, set fire to the wooden walls, though
no doubt they were too excited in the chase after you to observe it at the time. Now keep your eyes
open in this crowd for your friends of last night, though I very much fear that they are a good
hundred miles off by now."