125 Larch Walk
"That is unusual in a young girl."

"She is of a quiet nature. Besides, she is not so very young. She is four-and-twenty."

"This matter, from what you say, seems to have been a shock to her also."

"Terrible! She is even more affected than I."

"You have neither of you any doubt as to your son's guilt?"