174 Holmefield Avenue
"I knew that we should find a ventilator before ever we came to Stoke Moran."

"My dear Holmes!"

"Oh, yes, I did. You remember in her statement she said that her sister could smell Dr. Roylott's
cigar. Now, of course that suggested at once that there must be a communication between the two
rooms. It could only be a small one, or it would have been remarked upon at the coroner's inquiry. I
deduced a ventilator."

"But what harm can there be in that?"

"Well, there is at least a curious coincidence of dates. A ventilator is made, a cord is hung, and a
lady who sleeps in the bed dies. Does not that strike you?"