23 Wilson Gardens
Sherlock Holmes laughed. "Here is the foresight," said he putting his finger upon the little disc
and loop of the hat-securer.  "They are never sold upon hats. If this man ordered one, it is a sign
of a certain amount of foresight, since he went out of his way to take this precaution against the
wind. But since we see that he has broken the elastic and has not troubled to replace it, it is
obvious that he has less foresight now than formerly, which is a distinct proof of a weakening
nature. On the other hand, he has endeavoured to conceal some of these stains upon the felt by
daubing them with ink, which is a sign that he has not entirely lost his self-respect."

"Your reasoning is certainly plausible."

"The further points, that he is middle-aged, that his hair is grizzled, that it has been recently
cut, and that he uses lime-cream, are all to be gathered from a close examination of the lower part
of the lining. The lens discloses a large number of hair-ends, clean cut by the scissors of the
barber. They all appear to be adhesive, and there is a distinct odour of lime-cream. This dust, you
will observe, is not the gritty, grey dust of the street but the fluffy brown dust of the house,
showing that it has been hung up indoors most of the time, while the marks of moisture upon the
inside are proof positive that the wearer perspired very freely, and could therefore, hardly be in
the best of training."

"But his wife--you said that she had ceased to love him."

"This hat has not been brushed for weeks. When I see you, my dear Watson, with a week's accumulation
of dust upon your hat, and when your wife allows you to go out in such a state, I shall fear that
you also have been unfortunate enough to lose your wife's affection."