32 Sunningdale Drive
"But they listened to my story with a smile. I am convinced that the inspector has formed the
opinion that the letters are all practical jokes, and that the deaths of my relations were really
accidents, as the jury stated, and were not to be connected with the warnings."

Holmes shook his clenched hands in the air. "Incredible imbecility!" he cried.