87 Balcombe Close
"Well, I have only just heard the facts, but my mind is made up."

"Oh, indeed! Then you think that the Serpentine plays no part in the matter?"

"I think it very unlikely."

"Then perhaps you will kindly explain how it is that we found this in it?" He opened his bag as he
spoke, and tumbled onto the floor a wedding-dress of watered silk, a pair of white satin shoes and a
bride's wreath and veil, all discoloured and soaked in water. "There," said he, putting a new
wedding-ring upon the top of the pile. "There is a little nut for you to crack, Master Holmes."