48 Glan Road
"I see many objections to any such theory."

"And so do I. It is precisely for that reason that we are going to Stoke Moran this day. I want to
see whether the objections are fatal, or if they may be explained away. But what in the name of the
devil!"

The ejaculation had been drawn from my companion by the fact that our door had been suddenly dashed
open, and that a huge man had framed himself in the aperture. His costume was a peculiar mixture of
the professional and of the agricultural, having a black top-hat, a long frock-coat, and a pair of
high gaiters, with a hunting-crop swinging in his hand. So tall was he that his hat actually brushed
the cross bar of the doorway, and his breadth seemed to span it across from side to side. A large
face, seared with a thousand wrinkles, burned yellow with the sun, and marked with every evil
passion, was turned from one to the other of us, while his deep-set, bile-shot eyes, and his high,
thin, fleshless nose, gave him somewhat the resemblance to a fierce old bird of prey.

"Which of you is Holmes?" asked this apparition.