42 Harvey Street
"Oh, no, not now. I shall have to tell my tale to the police; but, between ourselves, if it were not
for the convincing evidence of this wound of mine, I should be surprised if they believed my
statement, for it is a very extraordinary one, and I have not much in the way of proof with which to
back it up; and, even if they believe me, the clues which I can give them are so vague that it is a
question whether justice will be done."

"Ha!" cried I, "if it is anything in the nature of a problem which you desire to see solved, I
should strongly recommend you to come to my friend, Mr. Sherlock Holmes, before you go to the
official police."

"Oh, I have heard of that fellow," answered my visitor, "and I should be very glad if he would take
the matter up, though of course I must use the official police as well. Would you give me an
introduction to him?"

"I'll do better. I'll take you round to him myself."