50 Ventnor Terrace
"Anyhow, I have grasped one fact which you seem to find it difficult to get hold of," replied
Lestrade with some warmth.

"And that is--"

"That McCarthy senior met his death from McCarthy junior and that all theories to the contrary are
the merest moonshine."

"Well, moonshine is a brighter thing than fog," said Holmes, laughing. "But I am very much mistaken
if this is not Hatherley Farm upon the left."