20 Cliff Place
"And when will you call?"

"At eight in the morning. She will not be up, so that we shall have a clear field. Besides, we must
be prompt, for this marriage may mean a complete change in her life and habits. I must wire to the
King without delay."

We had reached Baker Street and had stopped at the door. He was searching his pockets for the key
when someone passing said:

"Good-night, Mister Sherlock Holmes."

There were several people on the pavement at the time, but the greeting appeared to come from a slim
youth in an ulster who had hurried by.