103 Corbet Close
"'One of the most precious public possessions of the empire,' said I.

"'Precisely.' He opened the case, and there, imbedded in soft, flesh-coloured velvet, lay the
magnificent piece of jewellery which he had named. 'There are thirty-nine enormous beryls,' said he,
'and the price of the gold chasing is incalculable. The lowest estimate would put the worth of the
coronet at double the sum which I have asked. I am prepared to leave it with you as my security.'