65 Ventnor Terrace
"I know where it is."

"And how did you find out?"

"She showed me, as I told you she would."

"I am still in the dark."

"I do not wish to make a mystery," said he, laughing. "The matter was perfectly simple. You, of
course, saw that everyone in the street was an accomplice. They were all engaged for the evening."

"I guessed as much."