90 Balcombe Close
"We went upstairs together, the colonel first with the lamp, the fat manager and I behind him. It
was a labyrinth of an old house, with corridors, passages, narrow winding staircases, and little low
doors, the thresholds of which were hollowed out by the generations who had crossed them. There were
no carpets and no signs of any furniture above the ground floor, while the plaster was peeling off
the walls, and the damp was breaking through in green, unhealthy blotches. I tried to put on as
unconcerned an air as possible, but I had not forgotten the warnings of the lady, even though I
disregarded them, and I kept a keen eye upon my two companions. Ferguson appeared to be a morose and
silent man, but I could see from the little that he said that he was at least a fellow-countryman.

"Colonel Lysander Stark stopped at last before a low door, which he unlocked. Within was a small,
square room, in which the three of us could hardly get at one time. Ferguson remained outside, and
the colonel ushered me in.

"'We are now,' said he, 'actually within the hydraulic press, and it would be a particularly
unpleasant thing for us if anyone were to turn it on. The ceiling of this small chamber is really
the end of the descending piston, and it comes down with the force of many tons upon this metal
floor. There are small lateral columns of water outside which receive the force, and which transmit
and multiply it in the manner which is familiar to you. The machine goes readily enough, but there
is some stiffness in the working of it, and it has lost a little of its force. Perhaps you will have
the goodness to look it over and to show us how we can set it right.'

"I took the lamp from him, and I examined the machine very thoroughly. It was indeed a gigantic one,
and capable of exercising enormous pressure. When I passed outside, however, and pressed down the
levers which controlled it, I knew at once by the whishing sound that there was a slight leakage,
which allowed a regurgitation of water through one of the side cylinders. An examination showed that
one of the india-rubber bands which was round the head of a driving-rod had shrunk so as not quite
to fill the socket along which it worked. This was clearly the cause of the loss of power, and I
pointed it out to my companions, who followed my remarks very carefully and asked several practical
questions as to how they should proceed to set it right. When I had made it clear to them, I
returned to the main chamber of the machine and took a good look at it to satisfy my own curiosity.
It was obvious at a glance that the story of the fuller's-earth was the merest fabrication, for it
would be absurd to suppose that so powerful an engine could be designed for so inadequate a purpose.
The walls were of wood, but the floor consisted of a large iron trough, and when I came to examine
it I could see a crust of metallic deposit all over it. I had stooped and was scraping at this to
see exactly what it was when I heard a muttered exclamation in German and saw the cadaverous face of
the colonel looking down at me.