66 Holmefield Avenue
"Ah, yes, of course! Well, a cheetah is just a big cat, and yet a saucer of milk does not go very
far in satisfying its wants, I daresay. There is one point which I should wish to determine." He
squatted down in front of the wooden chair and examined the seat of it with the greatest attention.

"Thank you. That is quite settled," said he, rising and putting his lens in his pocket. "Hullo! Here
is something interesting!"