64 Bowler Street
We had the carriage to ourselves save for an immense litter of papers which Holmes had brought with
him. Among these he rummaged and read, with intervals of note-taking and of meditation, until we
were past Reading. Then he suddenly rolled them all into a gigantic ball and tossed them up onto the
rack.

"Have you heard anything of the case?" he asked.

"Not a word. I have not seen a paper for some days."

"The London press has not had very full accounts. I have just been looking through all the recent
papers in order to master the particulars. It seems, from what I gather, to be one of those simple
cases which are so extremely difficult."

"That sounds a little paradoxical."