74 Holmefield Avenue
"Ha! you find it so?"

"Extremely so. I congratulate you warmly."

Lestrade rose in his triumph and bent his head to look. "Why," he shrieked, "you're looking at the
wrong side!"

"On the contrary, this is the right side."

"The right side? You're mad! Here is the note written in pencil over here."

"And over here is what appears to be the fragment of a hotel bill, which interests me deeply."