32 Valentia Road
"No, no, my girl, the matter must be probed to the bottom."

"But I am so sure that he is innocent. You know what woman's instincts are. I know that he has done
no harm and that you will be sorry for having acted so harshly."

"Why is he silent, then, if he is innocent?"