160 Larch Walk
"It was Catherine Cusack who told me of it," said he in a crackling voice.

"I see--her ladyship's waiting-maid. Well, the temptation of sudden wealth so easily acquired was
too much for you, as it has been for better men before you; but you were not very scrupulous in the
means you used. It seems to me, Ryder, that there is the making of a very pretty villain in you. You
knew that this man Horner, the plumber, had been concerned in some such matter before, and that
suspicion would rest the more readily upon him. What did you do, then? You made some small job in my
lady's room--you and your confederate Cusack--and you managed that he should be the man sent for.
Then, when he had left, you rifled the jewel-case, raised the alarm, and had this unfortunate man
arrested. You then--"