113 Wantage Close
"I want to test a little theory of mine," said he, pulling on his boots. "I think, Watson, that you
are now standing in the presence of one of the most absolute fools in Europe. I deserve to be kicked
from here to Charing Cross. But I think I have the key of the affair now."

"And where is it?" I asked, smiling.

"In the bathroom," he answered. "Oh, yes, I am not joking," he continued, seeing my look of
incredulity. "I have just been there, and I have taken it out, and I have got it in this Gladstone
bag. Come on, my boy, and we shall see whether it will not fit the lock."

We made our way downstairs as quietly as possible, and out into the bright morning sunshine. In the
road stood our horse and trap, with the half-clad stable-boy waiting at the head. We both sprang in,
and away we dashed down the London Road. A few country carts were stirring, bearing in vegetables to
the metropolis, but the lines of villas on either side were as silent and lifeless as some city in a
dream.

"It has been in some points a singular case," said Holmes, flicking the horse on into a gallop. "I
confess that I have been as blind as a mole, but it is better to learn wisdom late than never to
learn it at all."