144 Holmefield Avenue
"And it is--"

"That Miss Flora Millar, the lady who had caused the disturbance, has actually been arrested. It
appears that she was formerly a danseuse at the Allegro, and that she has known the bridegroom for
some years. There are no further particulars, and the whole case is in your hands now--so far as it
has been set forth in the public press."