129 Larch Walk
"Have you never--" said Sherlock Holmes, bending forward and sinking his voice--"have you never
heard of the Ku Klux Klan?"

"I never have."

Holmes turned over the leaves of the book upon his knee. "Here it is," said he presently:

"'Ku Klux Klan. A name derived from the fanciful resemblance to the sound produced by cocking a
rifle. This terrible secret society was formed by some ex-Confederate soldiers in the Southern
states after the Civil War, and it rapidly formed local branches in different parts of the country,
notably in Tennessee, Louisiana, the Carolinas, Georgia, and Florida. Its power was used for
political purposes, principally for the terrorising of the negro voters and the murdering and
driving from the country of those who were opposed to its views. Its outrages were usually preceded
by a warning sent to the marked man in some fantastic but generally recognised shape--a sprig of
oak-leaves in some parts, melon seeds or orange pips in others. On receiving this the victim might
either openly abjure his former ways, or might fly from the country. If he braved the matter out,
death would unfailingly come upon him, and usually in some strange and unforeseen manner. So perfect
was the organisation of the society, and so systematic its methods, that there is hardly a case upon
record where any man succeeded in braving it with impunity, or in which any of its outrages were
traced home to the perpetrators. For some years the organisation flourished in spite of the efforts
of the United States government and of the better classes of the community in the South. Eventually,
in the year 1869, the movement rather suddenly collapsed, although there have been sporadic
outbreaks of the same sort since that date.'

"You will observe," said Holmes, laying down the volume, "that the sudden breaking up of the society
was coincident with the disappearance of Openshaw from America with their papers. It may well have
been cause and effect. It is no wonder that he and his family have some of the more implacable
spirits upon their track. You can understand that this register and diary may implicate some of the
first men in the South, and that there may be many who will not sleep easy at night until it is
recovered."