41 Cliff Place
"Now then, Mr. Cocksure," said the salesman, "I thought that I was out of geese, but before I finish
you'll find that there is still one left in my shop. You see this little book?"

"Well?"

"That's the list of the folk from whom I buy. D'you see? Well, then, here on this page are the
country folk, and the numbers after their names are where their accounts are in the big ledger. Now,
then! You see this other page in red ink? Well, that is a list of my town suppliers. Now, look at
that third name. Just read it out to me."