4 Glan Road
"So much so," I remarked, "that of the last six cases which I have added to my notes, three have
been entirely free of any legal crime."

"Precisely. You allude to my attempt to recover the Irene Adler papers, to the singular case of Miss
Mary Sutherland, and to the adventure of the man with the twisted lip. Well, I have no doubt that
this small matter will fall into the same innocent category. You know Peterson, the commissionaire?"