152 Beaconside
"'It is locked up,' I answered.

"'Oh, any old key will fit that bureau. When I was a youngster I have opened it myself with the key
of the box-room cupboard.'

"He often had a wild way of talking, so that I thought little of what he said. He followed me to my
room, however, that night with a very grave face.