44 Glan Road
"Never to return."

"And the papers?" asked the King hoarsely. "All is lost."

"We shall see." He pushed past the servant and rushed into the drawing-room, followed by the King
and myself. The furniture was scattered about in every direction, with dismantled shelves and open
drawers, as if the lady had hurriedly ransacked them before her flight. Holmes rushed at the bell-
pull, tore back a small sliding shutter, and, plunging in his hand, pulled out a photograph and a
letter. The photograph was of Irene Adler herself in evening dress, the letter was superscribed to
"Sherlock Holmes, Esq. To be left till called for." My friend tore it open and we all three read it
together. It was dated at midnight of the preceding night and ran in this way: