79 Badgers Dene
"And who is her maid?"

"Alice is her name. She is an American and came from California with her."

"A confidential servant?"

"A little too much so. It seemed to me that her mistress allowed her to take great liberties. Still,
of course, in America they look upon these things in a different way."