25 Beaconside
"Pending the alterations, as I understand. By the way, there does not seem to be any very pressing
need for repairs at that end wall."

"There were none. I believe that it was an excuse to move me from my room."

"Ah! that is suggestive. Now, on the other side of this narrow wing runs the corridor from which
these three rooms open. There are windows in it, of course?"

"Yes, but very small ones. Too narrow for anyone to pass through."