100 Beaconside
A heavily timbered park stretched up in a gentle slope, thickening into a grove at the highest
point. From amid the branches there jutted out the grey gables and high roof-tree of a very old
mansion.

"Stoke Moran?" said he.

"Yes, sir, that be the house of Dr. Grimesby Roylott," remarked the driver.

"There is some building going on there," said Holmes; "that is where we are going."