95 Corbet Close
"Really! Does it not strike you as a little singular that this McCarthy, who appears to have had
little of his own, and to have been under such obligations to Turner, should still talk of marrying
his son to Turner's daughter, who is, presumably, heiress to the estate, and that in such a very
cocksure manner, as if it were merely a case of a proposal and all else would follow? It is the more
strange, since we know that Turner himself was averse to the idea. The daughter told us as much. Do
you not deduce something from that?"

"We have got to the deductions and the inferences," said Lestrade, winking at me. "I find it hard
enough to tackle facts, Holmes, without flying away after theories and fancies."

"You are right," said Holmes demurely; "you do find it very hard to tackle the facts."