43 Wellswood Gardens
"At the same time," he remarked after a pause, during which he had sat puffing at his long pipe and
gazing down into the fire, "you can hardly be open to a charge of sensationalism, for out of these
cases which you have been so kind as to interest yourself in, a fair proportion do not treat of
crime, in its legal sense, at all. The small matter in which I endeavoured to help the King of
Bohemia, the singular experience of Miss Mary Sutherland, the problem connected with the man with
the twisted lip, and the incident of the noble bachelor, were all matters which are outside the pale
of the law. But in avoiding the sensational, I fear that you may have bordered on the trivial."

"The end may have been so," I answered, "but the methods I hold to have been novel and of interest."