24 Wantage Close
"Well, yes. Save, perhaps, that. And yet this John Openshaw seems to me to be walking amid even
greater perils than did the Sholtos."

"But have you," I asked, "formed any definite conception as to what these perils are?"

"There can be no question as to their nature," he answered.

"Then what are they? Who is this K. K. K., and why does he pursue this unhappy family?"

Sherlock Holmes closed his eyes and placed his elbows upon the arms of his chair, with his finger-
tips together. "The ideal reasoner," he remarked, "would, when he had once been shown a single fact
in all its bearings, deduce from it not only all the chain of events which led up to it but also all
the results which would follow from it. As Cuvier could correctly describe a whole animal by the
contemplation of a single bone, so the observer who has thoroughly understood one link in a series
of incidents should be able to accurately state all the other ones, both before and after. We have
not yet grasped the results which the reason alone can attain to. Problems may be solved in the
study which have baffled all those who have sought a solution by the aid of their senses. To carry
the art, however, to its highest pitch, it is necessary that the reasoner should be able to utilise
all the facts which have come to his knowledge; and this in itself implies, as you will readily see,
a possession of all knowledge, which, even in these days of free education and encyclopaedias, is a
somewhat rare accomplishment. It is not so impossible, however, that a man should possess all
knowledge which is likely to be useful to him in his work, and this I have endeavoured in my case to
do. If I remember rightly, you on one occasion, in the early days of our friendship, defined my
limits in a very precise fashion."