34 Larch Walk
"'Oh, yes,' said he, turning to me, 'we are very much obliged to you, Miss Hunter, for falling in
with our whims so far as to cut your hair. I assure you that it has not detracted in the tiniest
iota from your appearance. We shall now see how the electric-blue dress will become you. You will
find it laid out upon the bed in your room, and if you would be so good as to put it on we should
both be extremely obliged.'

"The dress which I found waiting for me was of a peculiar shade of blue. It was of excellent
material, a sort of beige, but it bore unmistakable signs of having been worn before. It could not
have been a better fit if I had been measured for it. Both Mr. and Mrs. Rucastle expressed a delight
at the look of it, which seemed quite exaggerated in its vehemence. They were waiting for me in the
drawing-room, which is a very large room, stretching along the entire front of the house, with three
long windows reaching down to the floor. A chair had been placed close to the central window, with
its back turned towards it. In this I was asked to sit, and then Mr. Rucastle, walking up and down
on the other side of the room, began to tell me a series of the funniest stories that I have ever
listened to. You cannot imagine how comical he was, and I laughed until I was quite weary. Mrs.
Rucastle, however, who has evidently no sense of humour, never so much as smiled, but sat with her
hands in her lap, and a sad, anxious look upon her face. After an hour or so, Mr. Rucastle suddenly
remarked that it was time to commence the duties of the day, and that I might change my dress and go
to little Edward in the nursery.

"Two days later this same performance was gone through under exactly similar circumstances. Again I
changed my dress, again I sat in the window, and again I laughed very heartily at the funny stories
of which my employer had an immense répertoire, and which he told inimitably. Then he handed me a
yellow-backed novel, and moving my chair a little sideways, that my own shadow might not fall upon
the page, he begged me to read aloud to him. I read for about ten minutes, beginning in the heart of
a chapter, and then suddenly, in the middle of a sentence, he ordered me to cease and to change my
dress.

"You can easily imagine, Mr. Holmes, how curious I became as to what the meaning of this
extraordinary performance could possibly be. They were always very careful, I observed, to turn my
face away from the window, so that I became consumed with the desire to see what was going on behind
my back. At first it seemed to be impossible, but I soon devised a means. My hand-mirror had been
broken, so a happy thought seized me, and I concealed a piece of the glass in my handkerchief. On
the next occasion, in the midst of my laughter, I put my handkerchief up to my eyes, and was able
with a little management to see all that there was behind me. I confess that I was disappointed.
There was nothing. At least that was my first impression. At the second glance, however, I perceived
that there was a man standing in the Southampton Road, a small bearded man in a grey suit, who
seemed to be looking in my direction. The road is an important highway, and there are usually people
there. This man, however, was leaning against the railings which bordered our field and was looking
earnestly up. I lowered my handkerchief and glanced at Mrs. Rucastle to find her eyes fixed upon me
with a most searching gaze. She said nothing, but I am convinced that she had divined that I had a
mirror in my hand and had seen what was behind me. She rose at once.

"'Jephro,' said she, 'there is an impertinent fellow upon the road there who stares up at Miss
Hunter.'