41 Down Close
"'Have you never heard of the League of the Red-headed Men?' he asked with his eyes open.

"'Never.'

"'Why, I wonder at that, for you are eligible yourself for one of the vacancies.'

"'And what are they worth?' I asked.

"'Oh, merely a couple of hundred a year, but the work is slight, and it need not interfere very much
with one's other occupations.'