137 Bowler Street
It was a quarter-past nine when I started from home and made my way across the Park, and so through
Oxford Street to Baker Street. Two hansoms were standing at the door, and as I entered the passage I
heard the sound of voices from above. On entering his room I found Holmes in animated conversation
with two men, one of whom I recognised as Peter Jones, the official police agent, while the other
was a long, thin, sad-faced man, with a very shiny hat and oppressively respectable frock-coat.

"Ha! Our party is complete," said Holmes, buttoning up his pea-jacket and taking his heavy hunting
crop from the rack. "Watson, I think you know Mr. Jones, of Scotland Yard? Let me introduce you to
Mr. Merryweather, who is to be our companion in to-night's adventure."

"We're hunting in couples again, Doctor, you see," said Jones in his consequential way. "Our friend
here is a wonderful man for starting a chase. All he wants is an old dog to help him to do the
running down."

"I hope a wild goose may not prove to be the end of our chase," observed Mr. Merryweather gloomily.

"You may place considerable confidence in Mr. Holmes, sir," said the police agent loftily. "He has
his own little methods, which are, if he won't mind my saying so, just a little too theoretical and
fantastic, but he has the makings of a detective in him. It is not too much to say that once or
twice, as in that business of the Sholto murder and the Agra treasure, he has been more nearly
correct than the official force."