116 Bowler Street
"'The work appears to be light and the pay munificent.'

"'Precisely so. We shall want you to come to-night by the last train.'

"'Where to?'

"'To Eyford, in Berkshire. It is a little place near the borders of Oxfordshire, and within seven
miles of Reading. There is a train from Paddington which would bring you there at about 11.15.'

"'Very good.'

"'I shall come down in a carriage to meet you.'