9 Cliff Place
Sherlock Holmes hailed a four-wheeler which was passing. "In that case we had better discuss it in a
cosy room rather than in this wind-swept market-place," said he. "But pray tell me, before we go
farther, who it is that I have the pleasure of assisting."

The man hesitated for an instant. "My name is John Robinson," he answered with a sidelong glance.

"No, no; the real name," said Holmes sweetly. "It is always awkward doing business with an alias."