122 Valentia Road
"That is our signal," said Holmes, springing to his feet; "it comes from the middle window."

As we passed out he exchanged a few words with the landlord, explaining that we were going on a late
visit to an acquaintance, and that it was possible that we might spend the night there. A moment
later we were out on the dark road, a chill wind blowing in our faces, and one yellow light
twinkling in front of us through the gloom to guide us on our sombre errand.

There was little difficulty in entering the grounds, for unrepaired breaches gaped in the old park
wall. Making our way among the trees, we reached the lawn, crossed it, and were about to enter
through the window when out from a clump of laurel bushes there darted what seemed to be a hideous
and distorted child, who threw itself upon the grass with writhing limbs and then ran swiftly across
the lawn into the darkness.

"My God!" I whispered; "did you see it?"

Holmes was for the moment as startled as I. His hand closed like a vice upon my wrist in his
agitation. Then he broke into a low laugh and put his lips to my ear.

"It is a nice household," he murmured. "That is the baboon."