36 Glendale
"Very. But will he see it?"

"Well, he is sure to keep an eye on the papers, since, to a poor man, the loss was a heavy one. He
was clearly so scared by his mischance in breaking the window and by the approach of Peterson that
he thought of nothing but flight, but since then he must have bitterly regretted the impulse which
caused him to drop his bird. Then, again, the introduction of his name will cause him to see it, for
everyone who knows him will direct his attention to it. Here you are, Peterson, run down to the
advertising agency and have this put in the evening papers."

"In which, sir?"

"Oh, in the Globe, Star, Pall Mall, St. James's, Evening News, Standard, Echo, and any others that
occur to you."

"Very well, sir. And this stone?"

"Ah, yes, I shall keep the stone. Thank you. And, I say, Peterson, just buy a goose on your way back
and leave it here with me, for we must have one to give to this gentleman in place of the one which
your family is now devouring."