162 Holmefield Avenue
"But he must be alive, Mr. Holmes."

"Unless this is a clever forgery to put us on the wrong scent. The ring, after all, proves nothing.
It may have been taken from him."

"No, no; it is, it is his very own writing!"

"Very well. It may, however, have been written on Monday and only posted to-day."

"That is possible."

"If so, much may have happened between."