107 Wantage Close
It was nearly four o'clock when we at last, after passing through the beautiful Stroud Valley, and
over the broad gleaming Severn, found ourselves at the pretty little country-town of Ross. A lean,
ferret-like man, furtive and sly-looking, was waiting for us upon the platform. In spite of the
light brown dustcoat and leather-leggings which he wore in deference to his rustic surroundings, I
had no difficulty in recognising Lestrade, of Scotland Yard. With him we drove to the Hereford Arms
where a room had already been engaged for us.

"I have ordered a carriage," said Lestrade as we sat over a cup of tea. "I knew your energetic
nature, and that you would not be happy until you had been on the scene of the crime."

"It was very nice and complimentary of you," Holmes answered. "It is entirely a question of
barometric pressure."

Lestrade looked startled. "I do not quite follow," he said.

"How is the glass? Twenty-nine, I see. No wind, and not a cloud in the sky. I have a caseful of
cigarettes here which need smoking, and the sofa is very much superior to the usual country hotel
abomination. I do not think that it is probable that I shall use the carriage to-night."