189 Beaconside
"Very naturally. By the way, about the bird, we were compelled to eat it."

"To eat it!" Our visitor half rose from his chair in his excitement.

"Yes, it would have been of no use to anyone had we not done so. But I presume that this other goose
upon the sideboard, which is about the same weight and perfectly fresh, will answer your purpose
equally well?"