70 Ventnor Terrace
"Yesterday morning I was seated in my office at the bank when a card was brought in to me by one of
the clerks. I started when I saw the name, for it was that of none other than--well, perhaps even to
you I had better say no more than that it was a name which is a household word all over the earth--
one of the highest, noblest, most exalted names in England. I was overwhelmed by the honour and
attempted, when he entered, to say so, but he plunged at once into business with the air of a man
who wishes to hurry quickly through a disagreeable task.

"'Mr. Holder,' said he, 'I have been informed that you are in the habit of advancing money.'

"'The firm does so when the security is good.' I answered.