133 Larch Walk
"What, sir! Oh, Heaven bless you!"

"No more words. Get out!"

And no more words were needed. There was a rush, a clatter upon the stairs, the bang of a door, and
the crisp rattle of running footfalls from the street.

"After all, Watson," said Holmes, reaching up his hand for his clay pipe, "I am not retained by the
police to supply their deficiencies. If Horner were in danger it would be another thing; but this
fellow will not appear against him, and the case must collapse. I suppose that I am commuting a
felony, but it is just possible that I am saving a soul. This fellow will not go wrong again; he is
too terribly frightened. Send him to jail now, and you make him a jail-bird for life. Besides, it is
the season of forgiveness. Chance has put in our way a most singular and whimsical problem, and its
solution is its own reward. If you will have the goodness to touch the bell, Doctor, we will begin
another investigation, in which, also a bird will be the chief feature."

On glancing over my notes of the seventy odd cases in which I have during the last eight years
studied the methods of my friend Sherlock Holmes, I find many tragic, some comic, a large number
merely strange, but none commonplace; for, working as he did rather for the love of his art than for
the acquirement of wealth, he refused to associate himself with any investigation which did not tend
towards the unusual, and even the fantastic. Of all these varied cases, however, I cannot recall any
which presented more singular features than that which was associated with the well-known Surrey
family of the Roylotts of Stoke Moran. The events in question occurred in the early days of my
association with Holmes, when we were sharing rooms as bachelors in Baker Street. It is possible
that I might have placed them upon record before, but a promise of secrecy was made at the time,
from which I have only been freed during the last month by the untimely death of the lady to whom
the pledge was given. It is perhaps as well that the facts should now come to light, for I have
reasons to know that there are widespread rumours as to the death of Dr. Grimesby Roylott which tend
to make the matter even more terrible than the truth.

It was early in April in the year '83 that I woke one morning to find Sherlock Holmes standing,
fully dressed, by the side of my bed. He was a late riser, as a rule, and as the clock on the
mantelpiece showed me that it was only a quarter-past seven, I blinked up at him in some surprise,
and perhaps just a little resentment, for I was myself regular in my habits.