17 Wilson Gardens
"Then they carried me in. She was bound to have me in. What else could she do? And into her sitting-
room, which was the very room which I suspected. It lay between that and her bedroom, and I was
determined to see which. They laid me on a couch, I motioned for air, they were compelled to open
the window, and you had your chance."

"How did that help you?"

"It was all-important. When a woman thinks that her house is on fire, her instinct is at once to
rush to the thing which she values most. It is a perfectly overpowering impulse, and I have more
than once taken advantage of it. In the case of the Darlington substitution scandal it was of use to
me, and also in the Arnsworth Castle business. A married woman grabs at her baby; an unmarried one
reaches for her jewel-box. Now it was clear to me that our lady of to-day had nothing in the house
more precious to her than what we are in quest of. She would rush to secure it. The alarm of fire
was admirably done. The smoke and shouting were enough to shake nerves of steel. She responded
beautifully. The photograph is in a recess behind a sliding panel just above the right bell-pull.
She was there in an instant, and I caught a glimpse of it as she half-drew it out. When I cried out
that it was a false alarm, she replaced it, glanced at the rocket, rushed from the room, and I have
not seen her since. I rose, and, making my excuses, escaped from the house. I hesitated whether to
attempt to secure the photograph at once; but the coachman had come in, and as he was watching me
narrowly it seemed safer to wait. A little over-precipitance may ruin all."