39 Balcombe Close
"I was very angry, for this was the third demand during the month. 'You shall not have a farthing
from me,' I cried, on which he bowed and left the room without another word.

"When he was gone I unlocked my bureau, made sure that my treasure was safe, and locked it again.
Then I started to go round the house to see that all was secure--a duty which I usually leave to
Mary but which I thought it well to perform myself that night. As I came down the stairs I saw Mary
herself at the side window of the hall, which she closed and fastened as I approached.