115 Holmefield Avenue
Never in my life have I seen such a sight. The man's face peeled off under the sponge like the bark
from a tree. Gone was the coarse brown tint! Gone, too, was the horrid scar which had seamed it
across, and the twisted lip which had given the repulsive sneer to the face! A twitch brought away
the tangled red hair, and there, sitting up in his bed, was a pale, sad-faced, refined-looking man,
black-haired and smooth-skinned, rubbing his eyes and staring about him with sleepy bewilderment.
Then suddenly realising the exposure, he broke into a scream and threw himself down with his face to
the pillow.

"Great heavens!" cried the inspector, "it is, indeed, the missing man. I know him from the
photograph."

The prisoner turned with the reckless air of a man who abandons himself to his destiny. "Be it so,"
said he. "And pray what am I charged with?"

"With making away with Mr. Neville St.--Oh, come, you can't be charged with that unless they make a
case of attempted suicide of it," said the inspector with a grin. "Well, I have been twenty-seven
years in the force, but this really takes the cake."

"If I am Mr. Neville St. Clair, then it is obvious that no crime has been committed, and that,
therefore, I am illegally detained."

"No crime, but a very great error has been committed," said Holmes. "You would have done better to
have trusted your wife."