10 Wellswood Gardens
"'You shall learn nothing from me,' said he with a passion such as I should not have thought was in
his nature. 'If you choose to call the police, let the police find what they can.'

"By this time the whole house was astir, for I had raised my voice in my anger. Mary was the first
to rush into my room, and, at the sight of the coronet and of Arthur's face, she read the whole
story and, with a scream, fell down senseless on the ground. I sent the house-maid for the police
and put the investigation into their hands at once. When the inspector and a constable entered the
house, Arthur, who had stood sullenly with his arms folded, asked me whether it was my intention to
charge him with theft. I answered that it had ceased to be a private matter, but had become a public
one, since the ruined coronet was national property. I was determined that the law should have its
way in everything.