10 Harvey Street
The fat man cast his eyes round, and then up at the open skylight.

"It is for me to ask you that," he shrieked, "you thieves! Spies and thieves! I have caught you,
have I? You are in my power. I'll serve you!" He turned and clattered down the stairs as hard as he
could go.

"He's gone for the dog!" cried Miss Hunter.

"I have my revolver," said I.

"Better close the front door," cried Holmes, and we all rushed down the stairs together. We had
hardly reached the hall when we heard the baying of a hound, and then a scream of agony, with a
horrible worrying sound which it was dreadful to listen to. An elderly man with a red face and
shaking limbs came staggering out at a side door.

"My God!" he cried. "Someone has loosed the dog. It's not been fed for two days. Quick, quick, or
it'll be too late!"