71 Wilson Gardens
"She was as bright as possible--at least until after the ceremony."

"And did you observe any change in her then?"

"Well, to tell the truth, I saw then the first signs that I had ever seen that her temper was just a
little sharp. The incident however, was too trivial to relate and can have no possible bearing upon
the case."

"Pray let us have it, for all that."

"Oh, it is childish. She dropped her bouquet as we went towards the vestry. She was passing the
front pew at the time, and it fell over into the pew. There was a moment's delay, but the gentleman
in the pew handed it up to her again, and it did not appear to be the worse for the fall. Yet when I
spoke to her of the matter, she answered me abruptly; and in the carriage, on our way home, she
seemed absurdly agitated over this trifling cause."

"Indeed! You say that there was a gentleman in the pew. Some of the general public were present,
then?"