82 Wellswood Gardens
"This Godfrey Norton was evidently an important factor in the matter. He was a lawyer. That sounded
ominous. What was the relation between them, and what the object of his repeated visits? Was she his
client, his friend, or his mistress? If the former, she had probably transferred the photograph to
his keeping. If the latter, it was less likely. On the issue of this question depended whether I
should continue my work at Briony Lodge, or turn my attention to the gentleman's chambers in the
Temple. It was a delicate point, and it widened the field of my inquiry. I fear that I bore you with
these details, but I have to let you see my little difficulties, if you are to understand the
situation."

"I am following you closely," I answered.