11 Wilson Gardens
About nine o'clock the light among the trees was extinguished, and all was dark in the direction of
the Manor House. Two hours passed slowly away, and then, suddenly, just at the stroke of eleven, a
single bright light shone out right in front of us.

"That is our signal," said Holmes, springing to his feet; "it comes from the middle window."