19 Harvey Street
"'I say a night's work, but an hour's would be nearer the mark. I simply want your opinion about a
hydraulic stamping machine which has got out of gear. If you show us what is wrong we shall soon set
it right ourselves. What do you think of such a commission as that?'

"'The work appears to be light and the pay munificent.'

"'Precisely so. We shall want you to come to-night by the last train.'

"'Where to?'

"'To Eyford, in Berkshire. It is a little place near the borders of Oxfordshire, and within seven
miles of Reading. There is a train from Paddington which would bring you there at about 11.15.'

"'Very good.'