10 Badgers Dene
"I thank your Majesty. Then there is no more to be done in the matter. I have the honour to wish you
a very good-morning." He bowed, and, turning away without observing the hand which the King had
stretched out to him, he set off in my company for his chambers.

And that was how a great scandal threatened to affect the kingdom of Bohemia, and how the best plans
of Mr. Sherlock Holmes were beaten by a woman's wit. He used to make merry over the cleverness of
women, but I have not heard him do it of late. And when he speaks of Irene Adler, or when he refers
to her photograph, it is always under the honourable title of the woman.

I had called upon my friend, Mr. Sherlock Holmes, one day in the autumn of last year and found him
in deep conversation with a very stout, florid-faced, elderly gentleman with fiery red hair. With an
apology for my intrusion, I was about to withdraw when Holmes pulled me abruptly into the room and
closed the door behind me.

"You could not possibly have come at a better time, my dear Watson," he said cordially.