85 Wantage Close
"Yes, sir. I am afraid that I am a little late, but I am not quite my own master, you know. I am
sorry that Miss Sutherland has troubled you about this little matter, for I think it is far better
not to wash linen of the sort in public. It was quite against my wishes that she came, but she is a
very excitable, impulsive girl, as you may have noticed, and she is not easily controlled when she
has made up her mind on a point. Of course, I did not mind you so much, as you are not connected
with the official police, but it is not pleasant to have a family misfortune like this noised
abroad. Besides, it is a useless expense, for how could you possibly find this Hosmer Angel?"

"On the contrary," said Holmes quietly; "I have every reason to believe that I will succeed in
discovering Mr. Hosmer Angel."

Mr. Windibank gave a violent start and dropped his gloves. "I am delighted to hear it," he said.

"It is a curious thing," remarked Holmes, "that a typewriter has really quite as much individuality
as a man's handwriting. Unless they are quite new, no two of them write exactly alike. Some letters
get more worn than others, and some wear only on one side. Now, you remark in this note of yours,
Mr. Windibank, that in every case there is some little slurring over of the 'e,' and a slight defect
in the tail of the 'r.' There are fourteen other characteristics, but those are the more obvious."