86 Wilson Gardens
"'And the work?'

"'Is to copy out the "Encyclopaedia Britannica." There is the first volume of it in that press. You
must find your own ink, pens, and blotting-paper, but we provide this table and chair. Will you be
ready to-morrow?'

"'Certainly,' I answered.

"'Then, good-bye, Mr. Jabez Wilson, and let me congratulate you once more on the important position
which you have been fortunate enough to gain.' He bowed me out of the room and I went home with my
assistant, hardly knowing what to say or do, I was so pleased at my own good fortune.