54 Badgers Dene
"I beg that you will not touch me with your filthy hands," remarked our prisoner as the handcuffs
clattered upon his wrists. "You may not be aware that I have royal blood in my veins. Have the
goodness, also, when you address me always to say 'sir' and 'please.'"

"All right," said Jones with a stare and a snigger. "Well, would you please, sir, march upstairs,
where we can get a cab to carry your Highness to the police-station?"

"That is better," said John Clay serenely. He made a sweeping bow to the three of us and walked
quietly off in the custody of the detective.