24 Harvey Street
"No; I shall be my own police. When I have spun the web they may take the flies, but not before."

All day I was engaged in my professional work, and it was late in the evening before I returned to
Baker Street. Sherlock Holmes had not come back yet. It was nearly ten o'clock before he entered,
looking pale and worn. He walked up to the sideboard, and tearing a piece from the loaf he devoured
it voraciously, washing it down with a long draught of water.