82 Holmefield Avenue
Our visitor staggered to his feet and clutched the mantelpiece with his right hand. Holmes unlocked
his strong-box and held up the blue carbuncle, which shone out like a star, with a cold, brilliant,
many-pointed radiance. Ryder stood glaring with a drawn face, uncertain whether to claim or to
disown it.

"The game's up, Ryder," said Holmes quietly. "Hold up, man, or you'll be into the fire! Give him an
arm back into his chair, Watson. He's not got blood enough to go in for felony with impunity. Give
him a dash of brandy. So! Now he looks a little more human. What a shrimp it is, to be sure!"

For a moment he had staggered and nearly fallen, but the brandy brought a tinge of colour into his
cheeks, and he sat staring with frightened eyes at his accuser.

"I have almost every link in my hands, and all the proofs which I could possibly need, so there is
little which you need tell me. Still, that little may as well be cleared up to make the case
complete. You had heard, Ryder, of this blue stone of the Countess of Morcar's?"

"It was Catherine Cusack who told me of it," said he in a crackling voice.

"I see--her ladyship's waiting-maid. Well, the temptation of sudden wealth so easily acquired was
too much for you, as it has been for better men before you; but you were not very scrupulous in the
means you used. It seems to me, Ryder, that there is the making of a very pretty villain in you. You
knew that this man Horner, the plumber, had been concerned in some such matter before, and that
suspicion would rest the more readily upon him. What did you do, then? You made some small job in my
lady's room--you and your confederate Cusack--and you managed that he should be the man sent for.
Then, when he had left, you rifled the jewel-case, raised the alarm, and had this unfortunate man
arrested. You then--"