29 Moor End
"You are the first who have ever heard my story. My father was a schoolmaster in Chesterfield, where
I received an excellent education. I travelled in my youth, took to the stage, and finally became a
reporter on an evening paper in London. One day my editor wished to have a series of articles upon
begging in the metropolis, and I volunteered to supply them. There was the point from which all my
adventures started. It was only by trying begging as an amateur that I could get the facts upon
which to base my articles. When an actor I had, of course, learned all the secrets of making up, and
had been famous in the green-room for my skill. I took advantage now of my attainments. I painted my
face, and to make myself as pitiable as possible I made a good scar and fixed one side of my lip in
a twist by the aid of a small slip of flesh-coloured plaster. Then with a red head of hair, and an
appropriate dress, I took my station in the business part of the city, ostensibly as a match-seller
but really as a beggar. For seven hours I plied my trade, and when I returned home in the evening I
found to my surprise that I had received no less than 26s. 4d.

"I wrote my articles and thought little more of the matter until, some time later, I backed a bill
for a friend and had a writ served upon me for £25. I was at my wit's end where to get the money,
but a sudden idea came to me. I begged a fortnight's grace from the creditor, asked for a holiday
from my employers, and spent the time in begging in the City under my disguise. In ten days I had
the money and had paid the debt.

"Well, you can imagine how hard it was to settle down to arduous work at £2 a week when I knew that
I could earn as much in a day by smearing my face with a little paint, laying my cap on the ground,
and sitting still. It was a long fight between my pride and the money, but the dollars won at last,
and I threw up reporting and sat day after day in the corner which I had first chosen, inspiring
pity by my ghastly face and filling my pockets with coppers. Only one man knew my secret. He was the
keeper of a low den in which I used to lodge in Swandam Lane, where I could every morning emerge as
a squalid beggar and in the evenings transform myself into a well-dressed man about town. This
fellow, a Lascar, was well paid by me for his rooms, so that I knew that my secret was safe in his
possession.

"Well, very soon I found that I was saving considerable sums of money. I do not mean that any beggar
in the streets of London could earn £700 a year--which is less than my average takings--but I had
exceptional advantages in my power of making up, and also in a facility of repartee, which improved
by practice and made me quite a recognised character in the City. All day a stream of pennies,
varied by silver, poured in upon me, and it was a very bad day in which I failed to take £2.

"As I grew richer I grew more ambitious, took a house in the country, and eventually married,
without anyone having a suspicion as to my real occupation. My dear wife knew that I had business in
the City. She little knew what.

"Last Monday I had finished for the day and was dressing in my room above the opium den when I
looked out of my window and saw, to my horror and astonishment, that my wife was standing in the
street, with her eyes fixed full upon me. I gave a cry of surprise, threw up my arms to cover my
face, and, rushing to my confidant, the Lascar, entreated him to prevent anyone from coming up to
me. I heard her voice downstairs, but I knew that she could not ascend. Swiftly I threw off my
clothes, pulled on those of a beggar, and put on my pigments and wig. Even a wife's eyes could not
pierce so complete a disguise. But then it occurred to me that there might be a search in the room,
and that the clothes might betray me. I threw open the window, reopening by my violence a small cut
which I had inflicted upon myself in the bedroom that morning. Then I seized my coat, which was
weighted by the coppers which I had just transferred to it from the leather bag in which I carried
my takings. I hurled it out of the window, and it disappeared into the Thames. The other clothes
would have followed, but at that moment there was a rush of constables up the stair, and a few
minutes after I found, rather, I confess, to my relief, that instead of being identified as Mr.
Neville St. Clair, I was arrested as his murderer.