112 Badgers Dene
"God bless you! You are doing what you can for him and for me. But it is too heavy a task. What was
he doing there at all? If his purpose were innocent, why did he not say so?"

"Precisely. And if it were guilty, why did he not invent a lie? His silence appears to me to cut
both ways. There are several singular points about the case. What did the police think of the noise
which awoke you from your sleep?"