68 Sunningdale Drive
It was an old rickety door and gave at once before our united strength. Together we rushed into the
room. It was empty. There was no furniture save a little pallet bed, a small table, and a basketful
of linen. The skylight above was open, and the prisoner gone.

"There has been some villainy here," said Holmes; "this beauty has guessed Miss Hunter's intentions
and has carried his victim off."

"But how?"

"Through the skylight. We shall soon see how he managed it." He swung himself up onto the roof. "Ah,
yes," he cried, "here's the end of a long light ladder against the eaves. That is how he did it."

"But it is impossible," said Miss Hunter; "the ladder was not there when the Rucastles went away."

"He has come back and done it. I tell you that he is a clever and dangerous man. I should not be
very much surprised if this were he whose step I hear now upon the stair. I think, Watson, that it
would be as well for you to have your pistol ready."