49 Glan Road
As he spoke there was a tap at the door, and the boy in buttons entered to announce Miss Mary
Sutherland, while the lady herself loomed behind his small black figure like a full-sailed merchant-
man behind a tiny pilot boat. Sherlock Holmes welcomed her with the easy courtesy for which he was
remarkable, and, having closed the door and bowed her into an armchair, he looked her over in the
minute and yet abstracted fashion which was peculiar to him.

"Do you not find," he said, "that with your short sight it is a little trying to do so much
typewriting?"