97 Corbet Close
"You see," remarked Holmes, as we paced to and fro in front of the house, "this marriage rather
simplifies matters. The photograph becomes a double-edged weapon now. The chances are that she would
be as averse to its being seen by Mr. Godfrey Norton, as our client is to its coming to the eyes of
his princess. Now the question is--Where are we to find the photograph?"

"Where, indeed?"

"It is most unlikely that she carries it about with her. It is cabinet size. Too large for easy
concealment about a woman's dress. She knows that the King is capable of having her waylaid and
searched. Two attempts of the sort have already been made. We may take it, then, that she does not
carry it about with her."