29 Larch Walk
"Their reasons for their conduct. But you shall have it all just as it occurred. When I came down,
Mr. Rucastle met me here and drove me in his dog-cart to the Copper Beeches. It is, as he said,
beautifully situated, but it is not beautiful in itself, for it is a large square block of a house,
whitewashed, but all stained and streaked with damp and bad weather. There are grounds round it,
woods on three sides, and on the fourth a field which slopes down to the Southampton highroad, which
curves past about a hundred yards from the front door. This ground in front belongs to the house,
but the woods all round are part of Lord Southerton's preserves. A clump of copper beeches
immediately in front of the hall door has given its name to the place.

"I was driven over by my employer, who was as amiable as ever, and was introduced by him that
evening to his wife and the child. There was no truth, Mr. Holmes, in the conjecture which seemed to
us to be probable in your rooms at Baker Street. Mrs. Rucastle is not mad. I found her to be a
silent, pale-faced woman, much younger than her husband, not more than thirty, I should think, while
he can hardly be less than forty-five. From their conversation I have gathered that they have been
married about seven years, that he was a widower, and that his only child by the first wife was the
daughter who has gone to Philadelphia. Mr. Rucastle told me in private that the reason why she had
left them was that she had an unreasoning aversion to her stepmother. As the daughter could not have
been less than twenty, I can quite imagine that her position must have been uncomfortable with her
father's young wife.