56 Holmefield Avenue
"I must really ask you to be a little more quiet!" said Holmes severely. "You have already
imperilled the whole success of our expedition. Might I beg that you would have the goodness to sit
down upon one of those boxes, and not to interfere?"

The solemn Mr. Merryweather perched himself upon a crate, with a very injured expression upon his
face, while Holmes fell upon his knees upon the floor and, with the lantern and a magnifying lens,
began to examine minutely the cracks between the stones. A few seconds sufficed to satisfy him, for
he sprang to his feet again and put his glass in his pocket.

"We have at least an hour before us," he remarked, "for they can hardly take any steps until the
good pawnbroker is safely in bed. Then they will not lose a minute, for the sooner they do their
work the longer time they will have for their escape. We are at present, Doctor--as no doubt you
have divined--in the cellar of the City branch of one of the principal London banks. Mr.
Merryweather is the chairman of directors, and he will explain to you that there are reasons why the
more daring criminals of London should take a considerable interest in this cellar at present."

"It is our French gold," whispered the director. "We have had several warnings that an attempt might
be made upon it."

"Your French gold?"

"Yes. We had occasion some months ago to strengthen our resources and borrowed for that purpose
30,000 napoleons from the Bank of France. It has become known that we have never had occasion to
unpack the money, and that it is still lying in our cellar. The crate upon which I sit contains
2,000 napoleons packed between layers of lead foil. Our reserve of bullion is much larger at present
than is usually kept in a single branch office, and the directors have had misgivings upon the
subject."