93 Larch Walk
"Mr. John Turner," cried the hotel waiter, opening the door of our sitting-room, and ushering in a
visitor.

The man who entered was a strange and impressive figure. His slow, limping step and bowed shoulders
gave the appearance of decrepitude, and yet his hard, deep-lined, craggy features, and his enormous
limbs showed that he was possessed of unusual strength of body and of character. His tangled beard,
grizzled hair, and outstanding, drooping eyebrows combined to give an air of dignity and power to
his appearance, but his face was of an ashen white, while his lips and the corners of his nostrils
were tinged with a shade of blue. It was clear to me at a glance that he was in the grip of some
deadly and chronic disease.

"Pray sit down on the sofa," said Holmes gently. "You had my note?"

"Yes, the lodge-keeper brought it up. You said that you wished to see me here to avoid scandal."

"I thought people would talk if I went to the Hall."

"And why did you wish to see me?" He looked across at my companion with despair in his weary eyes,
as though his question was already answered.