52 Valentia Road
"Very strange!" muttered Holmes, pulling at the rope. "There are one or two very singular points
about this room. For example, what a fool a builder must be to open a ventilator into another room,
when, with the same trouble, he might have communicated with the outside air!"

"That is also quite modern," said the lady.