109 Sunningdale Drive
"Then we had a talk as to what we should do, and Frank was all for openness, but I was so ashamed of
it all that I felt as if I should like to vanish away and never see any of them again--just sending
a line to pa, perhaps, to show him that I was alive. It was awful to me to think of all those lords
and ladies sitting round that breakfast-table and waiting for me to come back. So Frank took my
wedding-clothes and things and made a bundle of them, so that I should not be traced, and dropped
them away somewhere where no one could find them. It is likely that we should have gone on to Paris
to-morrow, only that this good gentleman, Mr. Holmes, came round to us this evening, though how he
found us is more than I can think, and he showed us very clearly and kindly that I was wrong and
that Frank was right, and that we should be putting ourselves in the wrong if we were so secret.
Then he offered to give us a chance of talking to Lord St. Simon alone, and so we came right away
round to his rooms at once. Now, Robert, you have heard it all, and I am very sorry if I have given
you pain, and I hope that you do not think very meanly of me."

Lord St. Simon had by no means relaxed his rigid attitude, but had listened with a frowning brow and
a compressed lip to this long narrative.

"Excuse me," he said, "but it is not my custom to discuss my most intimate personal affairs in this
public manner."

"Then you won't forgive me? You won't shake hands before I go?"