90 Corbet Close
"Yes. We had occasion some months ago to strengthen our resources and borrowed for that purpose
30,000 napoleons from the Bank of France. It has become known that we have never had occasion to
unpack the money, and that it is still lying in our cellar. The crate upon which I sit contains
2,000 napoleons packed between layers of lead foil. Our reserve of bullion is much larger at present
than is usually kept in a single branch office, and the directors have had misgivings upon the
subject."

"Which were very well justified," observed Holmes. "And now it is time that we arranged our little
plans. I expect that within an hour matters will come to a head. In the meantime Mr. Merryweather,
we must put the screen over that dark lantern."

"And sit in the dark?"

"I am afraid so. I had brought a pack of cards in my pocket, and I thought that, as we were a partie
carrée, you might have your rubber after all. But I see that the enemy's preparations have gone so
far that we cannot risk the presence of a light. And, first of all, we must choose our positions.
These are daring men, and though we shall take them at a disadvantage, they may do us some harm
unless we are careful. I shall stand behind this crate, and do you conceal yourselves behind those.
Then, when I flash a light upon them, close in swiftly. If they fire, Watson, have no compunction
about shooting them down."