122 Badgers Dene
"To the Red-headed League: On account of the bequest of the late Ezekiah Hopkins, of Lebanon,
Pennsylvania, U. S. A., there is now another vacancy open which entitles a member of the League to a
salary of £4 a week for purely nominal services. All red-headed men who are sound in body and mind
and above the age of twenty-one years, are eligible. Apply in person on Monday, at eleven o'clock,
to Duncan Ross, at the offices of the League, 7 Pope's Court, Fleet Street."

"What on earth does this mean?" I ejaculated after I had twice read over the extraordinary
announcement.

Holmes chuckled and wriggled in his chair, as was his habit when in high spirits. "It is a little
off the beaten track, isn't it?" said he. "And now, Mr. Wilson, off you go at scratch and tell us
all about yourself, your household, and the effect which this advertisement had upon your fortunes.
You will first make a note, Doctor, of the paper and the date."

"It is The Morning Chronicle of April 27, 1890. Just two months ago."

"Very good. Now, Mr. Wilson?"

"Well, it is just as I have been telling you, Mr. Sherlock Holmes," said Jabez Wilson, mopping his
forehead; "I have a small pawnbroker's business at Coburg Square, near the City. It's not a very
large affair, and of late years it has not done more than just give me a living. I used to be able
to keep two assistants, but now I only keep one; and I would have a job to pay him but that he is
willing to come for half wages so as to learn the business."