85 Corbet Close
"And is that all?"

"Only one little item in another of the morning papers, but it is a suggestive one."

"And it is--"

"That Miss Flora Millar, the lady who had caused the disturbance, has actually been arrested. It
appears that she was formerly a danseuse at the Allegro, and that she has known the bridegroom for
some years. There are no further particulars, and the whole case is in your hands now--so far as it
has been set forth in the public press."

"And an exceedingly interesting case it appears to be. I would not have missed it for worlds. But
there is a ring at the bell, Watson, and as the clock makes it a few minutes after four, I have no
doubt that this will prove to be our noble client. Do not dream of going, Watson, for I very much
prefer having a witness, if only as a check to my own memory."