29 Balcombe Close
"This, I take it, belongs to the room in which you used to sleep, the centre one to your sister's,
and the one next to the main building to Dr. Roylott's chamber?"

"Exactly so. But I am now sleeping in the middle one."

"Pending the alterations, as I understand. By the way, there does not seem to be any very pressing
need for repairs at that end wall."

"There were none. I believe that it was an excuse to move me from my room."