1 Harvey Street
"Let me introduce you," he shouted, "to Mr. Neville St. Clair, of Lee, in the county of Kent."

Never in my life have I seen such a sight. The man's face peeled off under the sponge like the bark
from a tree. Gone was the coarse brown tint! Gone, too, was the horrid scar which had seamed it
across, and the twisted lip which had given the repulsive sneer to the face! A twitch brought away
the tangled red hair, and there, sitting up in his bed, was a pale, sad-faced, refined-looking man,
black-haired and smooth-skinned, rubbing his eyes and staring about him with sleepy bewilderment.
Then suddenly realising the exposure, he broke into a scream and threw himself down with his face to
the pillow.

"Great heavens!" cried the inspector, "it is, indeed, the missing man. I know him from the
photograph."