67 Ventnor Terrace
"There is at least a presumption that the vessel in which the man or men are is a sailing-ship. It
looks as if they always send their singular warning or token before them when starting upon their
mission. You see how quickly the deed followed the sign when it came from Dundee. If they had come
from Pondicherry in a steamer they would have arrived almost as soon as their letter. But, as a
matter of fact, seven weeks elapsed. I think that those seven weeks represented the difference
between the mail-boat which brought the letter and the sailing vessel which brought the writer."

"It is possible."