93 Bowler Street
"Still, if I had married Lord St. Simon, of course I'd have done my duty by him. We can't command
our love, but we can our actions. I went to the altar with him with the intention to make him just
as good a wife as it was in me to be. But you may imagine what I felt when, just as I came to the
altar rails, I glanced back and saw Frank standing and looking at me out of the first pew. I thought
it was his ghost at first; but when I looked again there he was still, with a kind of question in
his eyes, as if to ask me whether I were glad or sorry to see him. I wonder I didn't drop. I know
that everything was turning round, and the words of the clergyman were just like the buzz of a bee
in my ear. I didn't know what to do. Should I stop the service and make a scene in the church? I
glanced at him again, and he seemed to know what I was thinking, for he raised his finger to his
lips to tell me to be still. Then I saw him scribble on a piece of paper, and I knew that he was
writing me a note. As I passed his pew on the way out I dropped my bouquet over to him, and he
slipped the note into my hand when he returned me the flowers. It was only a line asking me to join
him when he made the sign to me to do so. Of course I never doubted for a moment that my first duty
was now to him, and I determined to do just whatever he might direct.

"When I got back I told my maid, who had known him in California, and had always been his friend. I
ordered her to say nothing, but to get a few things packed and my ulster ready. I know I ought to
have spoken to Lord St. Simon, but it was dreadful hard before his mother and all those great
people. I just made up my mind to run away and explain afterwards. I hadn't been at the table ten
minutes before I saw Frank out of the window at the other side of the road. He beckoned to me and
then began walking into the Park. I slipped out, put on my things, and followed him. Some woman came
talking something or other about Lord St. Simon to me--seemed to me from the little I heard as if he
had a little secret of his own before marriage also--but I managed to get away from her and soon
overtook Frank. We got into a cab together, and away we drove to some lodgings he had taken in
Gordon Square, and that was my true wedding after all those years of waiting. Frank had been a
prisoner among the Apaches, had escaped, came on to 'Frisco, found that I had given him up for dead
and had gone to England, followed me there, and had come upon me at last on the very morning of my
second wedding."

"I saw it in a paper," explained the American. "It gave the name and the church but not where the
lady lived."

"Then we had a talk as to what we should do, and Frank was all for openness, but I was so ashamed of
it all that I felt as if I should like to vanish away and never see any of them again--just sending
a line to pa, perhaps, to show him that I was alive. It was awful to me to think of all those lords
and ladies sitting round that breakfast-table and waiting for me to come back. So Frank took my
wedding-clothes and things and made a bundle of them, so that I should not be traced, and dropped
them away somewhere where no one could find them. It is likely that we should have gone on to Paris
to-morrow, only that this good gentleman, Mr. Holmes, came round to us this evening, though how he
found us is more than I can think, and he showed us very clearly and kindly that I was wrong and
that Frank was right, and that we should be putting ourselves in the wrong if we were so secret.
Then he offered to give us a chance of talking to Lord St. Simon alone, and so we came right away
round to his rooms at once. Now, Robert, you have heard it all, and I am very sorry if I have given
you pain, and I hope that you do not think very meanly of me."

Lord St. Simon had by no means relaxed his rigid attitude, but had listened with a frowning brow and
a compressed lip to this long narrative.

"Excuse me," he said, "but it is not my custom to discuss my most intimate personal affairs in this
public manner."