29 Bowler Street
"I searched the Dundee records, and when I found that the barque Lone Star was there in January,
'85, my suspicion became a certainty. I then inquired as to the vessels which lay at present in the
port of London."

"Yes?"

"The Lone Star had arrived here last week. I went down to the Albert Dock and found that she had
been taken down the river by the early tide this morning, homeward bound to Savannah. I wired to
Gravesend and learned that she had passed some time ago, and as the wind is easterly I have no doubt
that she is now past the Goodwins and not very far from the Isle of Wight."

"What will you do, then?"

"Oh, I have my hand upon him. He and the two mates, are as I learn, the only native-born Americans
in the ship. The others are Finns and Germans. I know, also, that they were all three away from the
ship last night. I had it from the stevedore who has been loading their cargo. By the time that
their sailing-ship reaches Savannah the mail-boat will have carried this letter, and the cable will
have informed the police of Savannah that these three gentlemen are badly wanted here upon a charge
of murder."

There is ever a flaw, however, in the best laid of human plans, and the murderers of John Openshaw
were never to receive the orange pips which would show them that another, as cunning and as resolute
as themselves, was upon their track. Very long and very severe were the equinoctial gales that year.
We waited long for news of the Lone Star of Savannah, but none ever reached us. We did at last hear
that somewhere far out in the Atlantic a shattered stern-post of a boat was seen swinging in the
trough of a wave, with the letters "L. S." carved upon it, and that is all which we shall ever know
of the fate of the Lone Star.