30 Corbet Close
"My dear fellow, I would not miss it for anything."

I had no keener pleasure than in following Holmes in his professional investigations, and in
admiring the rapid deductions, as swift as intuitions, and yet always founded on a logical basis
with which he unravelled the problems which were submitted to him. I rapidly threw on my clothes and
was ready in a few minutes to accompany my friend down to the sitting-room. A lady dressed in black
and heavily veiled, who had been sitting in the window, rose as we entered.