50 Wellswood Gardens
I placed my revolver, cocked, upon the top of the wooden case behind which I crouched. Holmes shot
the slide across the front of his lantern and left us in pitch darkness--such an absolute darkness
as I have never before experienced. The smell of hot metal remained to assure us that the light was
still there, ready to flash out at a moment's notice. To me, with my nerves worked up to a pitch of
expectancy, there was something depressing and subduing in the sudden gloom, and in the cold dank
air of the vault.

"They have but one retreat," whispered Holmes. "That is back through the house into Saxe-Coburg
Square. I hope that you have done what I asked you, Jones?"

"I have an inspector and two officers waiting at the front door."

"Then we have stopped all the holes. And now we must be silent and wait."

What a time it seemed! From comparing notes afterwards it was but an hour and a quarter, yet it
appeared to me that the night must have almost gone and the dawn be breaking above us. My limbs were
weary and stiff, for I feared to change my position; yet my nerves were worked up to the highest
pitch of tension, and my hearing was so acute that I could not only hear the gentle breathing of my
companions, but I could distinguish the deeper, heavier in-breath of the bulky Jones from the thin,
sighing note of the bank director. From my position I could look over the case in the direction of
the floor. Suddenly my eyes caught the glint of a light.