42 Holmefield Avenue
The man sat huddled up in his chair, with his head sunk upon his breast, like one who is utterly
crushed. Holmes stuck his feet up on the corner of the mantelpiece and, leaning back with his hands
in his pockets, began talking, rather to himself, as it seemed, than to us.

"The man married a woman very much older than himself for her money," said he, "and he enjoyed the
use of the money of the daughter as long as she lived with them. It was a considerable sum, for
people in their position, and the loss of it would have made a serious difference. It was worth an
effort to preserve it. The daughter was of a good, amiable disposition, but affectionate and warm-
hearted in her ways, so that it was evident that with her fair personal advantages, and her little
income, she would not be allowed to remain single long. Now her marriage would mean, of course, the
loss of a hundred a year, so what does her stepfather do to prevent it? He takes the obvious course
of keeping her at home and forbidding her to seek the company of people of her own age. But soon he
found that that would not answer forever. She became restive, insisted upon her rights, and finally
announced her positive intention of going to a certain ball. What does her clever stepfather do
then? He conceives an idea more creditable to his head than to his heart. With the connivance and
assistance of his wife he disguised himself, covered those keen eyes with tinted glasses, masked the
face with a moustache and a pair of bushy whiskers, sunk that clear voice into an insinuating
whisper, and doubly secure on account of the girl's short sight, he appears as Mr. Hosmer Angel, and
keeps off other lovers by making love himself."

"It was only a joke at first," groaned our visitor. "We never thought that she would have been so
carried away."

"Very likely not. However that may be, the young lady was very decidedly carried away, and, having
quite made up her mind that her stepfather was in France, the suspicion of treachery never for an
instant entered her mind. She was flattered by the gentleman's attentions, and the effect was
increased by the loudly expressed admiration of her mother. Then Mr. Angel began to call, for it was
obvious that the matter should be pushed as far as it would go if a real effect were to be produced.
There were meetings, and an engagement, which would finally secure the girl's affections from
turning towards anyone else. But the deception could not be kept up forever. These pretended
journeys to France were rather cumbrous. The thing to do was clearly to bring the business to an end
in such a dramatic manner that it would leave a permanent impression upon the young lady's mind and
prevent her from looking upon any other suitor for some time to come. Hence those vows of fidelity
exacted upon a Testament, and hence also the allusions to a possibility of something happening on
the very morning of the wedding. James Windibank wished Miss Sutherland to be so bound to Hosmer
Angel, and so uncertain as to his fate, that for ten years to come, at any rate, she would not
listen to another man. As far as the church door he brought her, and then, as he could go no
farther, he conveniently vanished away by the old trick of stepping in at one door of a four-wheeler
and out at the other. I think that was the chain of events, Mr. Windibank!"