43 Down Close
"'So,' said he, smiling, 'it was you, then. I thought that it must be when I saw the door open.'

"'Oh, I am so frightened!' I panted.

"'My dear young lady! my dear young lady!'--you cannot think how caressing and soothing his manner
was--'and what has frightened you, my dear young lady?'

"But his voice was just a little too coaxing. He overdid it. I was keenly on my guard against him.

"'I was foolish enough to go into the empty wing,' I answered. 'But it is so lonely and eerie in
this dim light that I was frightened and ran out again. Oh, it is so dreadfully still in there!'

"'Only that?' said he, looking at me keenly.