134 Wantage Close
"What do you make of that, Watson?"

"It's a common enough lash. But I don't know why it should be tied."

"That is not quite so common, is it? Ah, me! it's a wicked world, and when a clever man turns his
brains to crime it is the worst of all. I think that I have seen enough now, Miss Stoner, and with
your permission we shall walk out upon the lawn."