Crime scene report #384205
"You have a maid who has a sweetheart? I think that you remarked to your uncle last night that she
had been out to see him?"

"Yes, and she was the girl who waited in the drawing-room, and who may have heard uncle's remarks
about the coronet."

"I see. You infer that she may have gone out to tell her sweetheart, and that the two may have
planned the robbery."

"But what is the good of all these vague theories," cried the banker impatiently, "when I have told
you that I saw Arthur with the coronet in his hands?"

"Wait a little, Mr. Holder. We must come back to that. About this girl, Miss Holder. You saw her
return by the kitchen door, I presume?"

"Yes; when I went to see if the door was fastened for the night I met her slipping in. I saw the
man, too, in the gloom."