Crime scene report #577834
"Well, Watson," said Holmes when our visitor had left us, "what do you make of it all?"

"I make nothing of it," I answered frankly. "It is a most mysterious business."

"As a rule," said Holmes, "the more bizarre a thing is the less mysterious it proves to be. It is
your commonplace, featureless crimes which are really puzzling, just as a commonplace face is the
most difficult to identify. But I must be prompt over this matter."