Crime scene report #993884
"This went on day after day, Mr. Holmes, and on Saturday the manager came in and planked down four
golden sovereigns for my week's work. It was the same next week, and the same the week after. Every
morning I was there at ten, and every afternoon I left at two. By degrees Mr. Duncan Ross took to
coming in only once of a morning, and then, after a time, he did not come in at all. Still, of
course, I never dared to leave the room for an instant, for I was not sure when he might come, and
the billet was such a good one, and suited me so well, that I would not risk the loss of it.

"Eight weeks passed away like this, and I had written about Abbots and Archery and Armour and
Architecture and Attica, and hoped with diligence that I might get on to the B's before very long.
It cost me something in foolscap, and I had pretty nearly filled a shelf with my writings. And then
suddenly the whole business came to an end."

"To an end?"

"Yes, sir. And no later than this morning. I went to my work as usual at ten o'clock, but the door
was shut and locked, with a little square of cardboard hammered on to the middle of the panel with a
tack. Here it is, and you can read for yourself."

He held up a piece of white cardboard about the size of a sheet of note-paper. It read in this
fashion:

The Red-headed League          is       Dissolved   October 9, 1890.